New and improved documentation
The CSS is from the Haiku User Guide, the general layout as I've
used them for other docs before.

Not much could be copied from the original ReadMe.html, other
than the section about the "Add file extension" under the
screenshot of the "Renaming Modes". And even that I have heavily
edited to make it more concise.

Moved the original documentation "original_docs" in case
translators do want to take a look.

I used JavaScript to change the image of the examples for the
different renaming modes with a click.
Hope that's OK for everyone. There a no-script alternative.

All images that may be localized are moved in their language-coded
subfolder, like "images/en/" for English. Non-localized images,
like icons, stay in the "images" top level.

Feedback or improvements welcome!