Implement menu to show the ReadMe
Using BPathFinder to look into all B_FIND_PATH_DOCUMENTATION_DIRECTORY
locations for the documentation. When packaged, it'll be in e.g.
/system/documentation/packages/pecorename/

If you build yourself, put the ReadMe.html and images folder e.g. into
/system/non-packaged/documentation/packages/pecorename/

The file name of the ReadMe is set with the localization. If the translator
decides to call the German version "LiesMich.html", the German file
has to be name like that. It doesn't matter, really, but I'd prefer
"ReadMe_de.html".

Updated catkeys, only German at 100%.

Fix missing comma in ReadMe.html.